Column test: Create data and run tests.
Added more to the initial work to be able to create and populate a
single Postgres table with one table.  The table contains a primary
key called 'id' and a number of columns determined by the test.  The
columns are named 'col000' through 'colNNN' (max 1000 columns).

The same work is done to create Apache Parquet tables using the Conda
installation kit, which has binary support for Parquet files.